+++ Extensions/IISWebAppDeploy/Tests/Tasks/SqlDacpacDeploy/SqlQueryOnTargetMachines.Tests.ps1
@@ -4,6 +4,7 @@ $sut = (Split-Path -Leaf $MyInvocation.MyCommand.Path).Replace(".Tests.", ".")
 $VerbosePreference = 'Continue'
 
 $sqlQueryOnTargetMachinesPath = "$currentScriptPath\..\..\..\Src\Tasks\$scriptDirName\TaskModuleSqlUtility\$sut"
+$sqlUtility = "$currentScriptPath\..\..\..\Src\Tasks\$scriptDirName\TaskModuleSqlUtility\sqlUtility.ps1"
 
 if(-not (Test-Path -Path $sqlQueryOnTargetMachinesPath ))
 {
@@ -11,6 +12,7 @@ if(-not (Test-Path -Path $sqlQueryOnTargetMachinesPath ))
 }
 
 . "$sqlQueryOnTargetMachinesPath"
+. "$sqlUtility"
 
 # Tests ----------------------------------------------------------------------------